Cleanup useTemplateApi + useTemplateContext
