+++ app/(gcforms)/[locale]/(form administration)/form-builder/[id]/settings/page.tsx
@@ -1,6 +1,6 @@
 import { serverTranslation } from "@i18n";
-import { ResponseDelivery } from "./components/ResponseDelivery";
 import { Metadata } from "next";
+import { ResponseDelivery } from "./components/ResponseDelivery";
 
 export async function generateMetadata({
   params: { locale },
